Move property update code from crafting vm into Item; use it in import
