+++ WPFSKillTree/ViewModels/Crafting/AbstractCraftingViewModel.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
-using System.Text.RegularExpressions;
 using EnumsNET;
 using PoESkillTree.Common.ViewModels;
 using PoESkillTree.Engine.GameModel.Items;
@@ -334,15 +332,7 @@ namespace PoESkillTree.ViewModels.Crafting
             var quality = SelectedBase.CanHaveQuality 
                 ? _qualitySlider.Value
                 : 0;
-            var properties = Item.BaseType.GetRawProperties(quality)
-                .Concat(GetAdditionalProperties());
-            Item.Properties = new ObservableCollection<ItemMod>(properties);
-            ApplyLocals();
-
-            if (Item.IsWeapon)
-            {
-                ApplyElementalMods(Item.Mods);
-            }
+            Item.UpdateProperties(quality, GetAdditionalProperties().ToArray());
 
             if (MsImplicits.Any())
             {
@@ -394,114 +384,6 @@ namespace PoESkillTree.ViewModels.Crafting
                 var attr = ItemMod.Numberfilter.Replace(line, "#");
                 var isLocal = ModifierLocalityTester.IsLocal(attr, SelectedBase.Tags);
                 yield return new ItemMod(line, isLocal);
-            }
-        }
-
-        private void ApplyElementalMods(IEnumerable<ItemMod> allMods)
-        {
-            var elementalMods = new List<ItemMod>();
-            var chaosMods = new List<ItemMod>();
-            foreach (var mod in allMods)
-            {
-                string attr = mod.Attribute;
-                if (attr.StartsWith("Adds") && !attr.Contains("in Main Hand") && !attr.Contains("in Off Hand"))
-                {
-                    if (attr.Contains("Fire") || attr.Contains("Cold") || attr.Contains("Lightning"))
-                    {
-                        elementalMods.Add(mod);
-                    }
-                    if (attr.Contains("Chaos"))
-                    {
-                        chaosMods.Add(mod);
-                    }
-                }
-            }
-
-            if (elementalMods.Any())
-            {
-                var values = new List<float>();
-                var mods = new List<string>();
-                var cols = new List<ValueColoring>();
-
-                var fmod = elementalMods.FirstOrDefault(m => m.Attribute.Contains("Fire"));
-                if (fmod != null)
-                {
-                    values.AddRange(fmod.Values);
-                    mods.Add("#-#");
-                    cols.Add(ValueColoring.Fire);
-                    cols.Add(ValueColoring.Fire);
-                }
-
-                var cmod = elementalMods.FirstOrDefault(m => m.Attribute.Contains("Cold"));
-                if (cmod != null)
-                {
-                    values.AddRange(cmod.Values);
-                    mods.Add("#-#");
-                    cols.Add(ValueColoring.Cold);
-                    cols.Add(ValueColoring.Cold);
-                }
-
-                var lmod = elementalMods.FirstOrDefault(m => m.Attribute.Contains("Lightning"));
-                if (lmod != null)
-                {
-                    values.AddRange(lmod.Values);
-                    mods.Add("#-#");
-                    cols.Add(ValueColoring.Lightning);
-                    cols.Add(ValueColoring.Lightning);
-                }
-
-                Item.Properties.Add(new ItemMod("Elemental Damage: " + string.Join(", ", mods), true, values, cols));
-            }
-
-            if (chaosMods.Any())
-            {
-                Item.Properties.Add(new ItemMod("Chaos Damage: #-#", true, chaosMods[0].Values,
-                    new[] { ValueColoring.Chaos, ValueColoring.Chaos }));
-            }
-        }
-
-        private void ApplyLocals()
-        {
-            foreach (var pair in Item.GetModsAffectingProperties())
-            {
-                ItemMod prop = pair.Key;
-                List<ItemMod> applymods = pair.Value;
-
-                List<ItemMod> percm = applymods.Where(m => Regex.IsMatch(m.Attribute, @"(?<!\+)#%")).ToList();
-                List<ItemMod> valuem = applymods.Except(percm).ToList();
-
-                if (valuem.Count > 0)
-                {
-                    IReadOnlyList<float> val = valuem
-                        .Select(m => m.Values)
-                        .Aggregate((l1, l2) => l1.Zip(l2, (f1, f2) => f1 + f2)
-                        .ToList());
-                    IReadOnlyList<float> nval = prop.Values
-                        .Zip(val, (f1, f2) => f1 + f2)
-                        .ToList();
-                    prop.ValueColors = prop.ValueColors
-                        .Select((c, i) => val[i] == nval[i] ? prop.ValueColors[i] : ValueColoring.LocallyAffected)
-                        .ToList();
-                    prop.Values = nval;
-                }
-
-                Func<float, float> roundf = val => (float)Math.Round(val);
-
-                if (prop.Attribute.Contains("Critical"))
-                {
-                    roundf = f => (float)(Math.Round(f * 10) / 10);
-                }
-                else if (prop.Attribute.Contains("per Second"))
-                {
-                    roundf = f => (float)(Math.Round(f * 100) / 100);
-                }
-
-                if (percm.Count > 0)
-                {
-                    var perc = 1f + percm.Select(m => m.Values[0]).Sum() / 100f;
-                    prop.ValueColors = prop.ValueColors.Select(c => ValueColoring.LocallyAffected).ToList();
-                    prop.Values = prop.Values.Select(v => roundf(v * perc)).ToList();
-                }
             }
         }
